requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
google-genai
requests==2.31.0
spacy==3.7.2
torch==2.1.0
transformers==4.44.0
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.0
newspaper3k==0.2.8
nltk==3.8.1
dbias==0.1.3
lxml==5.2.1
lxml_html_clean==0.1.1
google-cloud-aiplatform
gunicorn
spacytextblob
google-cloud-storage
tensorflow
tf-keras 
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
